Begin implementing ability to handle two builds
IMPORTANT: This has not been tested, and I have given it only a cursory
review. This is very much a work in progress.

I'm setting this aside for now, because after talking with Mariana, we may try
to accomplish this via two cases - since it's so messy to try to handle this in
a single case.

In addition to possibly needing to rework some of the stuff implemented here,
this is also missing implementation for these three functions:

    save_xml_file
    stage_saved_xml_file
    restore_or_save_xml_file

I envisioned that those would go in a utility module similar to user_nl_utils,
which would be wrapped in extensive unit tests.

Test suite: None
Test baseline: N/A
Test namelist changes: N/A
Test status: N/A

Fixes: None

User interface changes?: No

Code review: None