Do not use normalization when mapping glc2ocn
In talking this through with Mariana Vertenstein and Tony Craig, we came
to the conclusion that normalization should never be used when doing
runoff mapping (i.e., nearest neighbor with optional smoothing): doing
normalization breaks conservation. I have verified that the remapping is
conservative with norm=.false., but was not conservative with
norm=.true.